nna-app/nna-app/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic[email]==2.5.3
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
aiofiles==23.2.1
python-magic==0.4.27
